+++ tennisclub-api/src/main/java/com/frenchies/tennisclub/dto/OrderItemDTO.java
@@ -1,0 +1,83 @@
+package com.frenchies.tennisclub.dto;
+
+
+
+public class OrderItemDTO
+{
+    private Long id;
+
+    private ProductDTO product;
+
+    private Integer amount;
+
+    private NewPriceDTO pricePerItem;
+
+    public NewPriceDTO getPricePerItem() {
+        return pricePerItem;
+    }
+
+    public void setPricePerItem(NewPriceDTO pricePerItem) {
+        this.pricePerItem = pricePerItem;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public ProductDTO getProduct() {
+        return product;
+    }
+
+    public void setProduct(ProductDTO product) {
+        this.product = product;
+    }
+
+    public Integer getAmount() {
+        return amount;
+    }
+
+    public void setAmount(Integer amount) {
+        this.amount = amount;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((amount == null) ? 0 : amount.hashCode());
+        result = prime * result + ((id == null) ? 0 : id.hashCode());
+        result = prime * result + ((product == null) ? 0 : product.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        OrderItemDTO other = (OrderItemDTO) obj;
+        if (amount == null) {
+            if (other.amount != null)
+                return false;
+        } else if (!amount.equals(other.amount))
+            return false;
+        if (id == null) {
+            if (other.id != null)
+                return false;
+        } else if (!id.equals(other.id))
+            return false;
+        if (product == null) {
+            if (other.product != null)
+                return false;
+        } else if (!product.equals(other.product))
+            return false;
+        return true;
+    }
+}
