+++ tennisclub-api/src/main/java/com/frenchies/tennisclub/dto/UserAuthenticateDTO.java
@@ -1,0 +1,27 @@
+package com.frenchies.tennisclub.dto;
+
+public class UserAuthenticateDTO
+{
+    private Long userId;
+    private String password;
+
+    public Long getUserId()
+    {
+        return userId;
+    }
+
+    public void setUserId(Long userId)
+    {
+        this.userId = userId;
+    }
+
+    public String getPassword()
+    {
+        return password;
+    }
+
+    public void setPassword(String password)
+    {
+        this.password = password;
+    }
+}
