+++ tennisclub-api/src/main/java/com/frenchies/tennisclub/facade/CourtFacade.java
@@ -1,0 +1,9 @@
+package com.frenchies.tennisclub.facade;
+
+import java.util.List;
+
+public class CourtFacade {
+	
+	List<CourtDTO> getAllCourts();
+
+}
