--- tennisclub-api/src/main/java/cz/fi/muni/pa165/dto/UserAuthenticateDTO.java
@@ -1,27 +1,0 @@
-package cz.fi.muni.pa165.dto;
-
-public class UserAuthenticateDTO
-{
-    private Long userId;
-    private String password;
-
-    public Long getUserId()
-    {
-        return userId;
-    }
-
-    public void setUserId(Long userId)
-    {
-        this.userId = userId;
-    }
-
-    public String getPassword()
-    {
-        return password;
-    }
-
-    public void setPassword(String password)
-    {
-        this.password = password;
-    }
-}
