--- tennisclub-api/src/main/java/cz/fi/muni/pa165/facade/BookingFacade.java
@@ -1,12 +1,0 @@
-package cz.fi.muni.pa165.facade;
-
-import java.util.List;
-
-public class BookingFacade {
-	
-	List<BookingDTO> findAllBooking();
-
-	void createBooking(long idBooking, long idCourt, long idPlayer1, long idPlayer2, date dateOfBooking, Hour24 hourOfBooking);
-	
-	void deleteBooking(long idBooking)
-}
