--- tennisclub-api/src/main/java/cz/fi/muni/pa165/facade/CategoryFacade.java
@@ -1,13 +1,0 @@
-package cz.fi.muni.pa165.facade;
-
-import java.util.List;
-
-import cz.fi.muni.pa165.dto.CategoryCreateDTO;
-import cz.fi.muni.pa165.dto.CategoryDTO;
-public interface CategoryFacade
-{
-    List<CategoryDTO> getAllCategories();
-    CategoryDTO getCategoryById(Long id);
-
-    Long createCategory(CategoryCreateDTO categoryCreateDTO);
-}
