--- tennisclub-api/src/main/java/cz/fi/muni/pa165/facade/OrderFacade.java
@@ -1,17 +1,0 @@
-package cz.fi.muni.pa165.facade;
-
-import java.util.List;
-
-import cz.fi.muni.pa165.dto.OrderDTO;
-import cz.fi.muni.pa165.enums.OrderState;
-
-public interface OrderFacade {
-	public List<OrderDTO> getAllOrders();
-	public List<OrderDTO> getAllOrdersLastWeek(OrderState state);
-	public List<OrderDTO> getOrdersByUser(Long userId);
-	public List<OrderDTO> getOrdersByState(OrderState state);
-	public OrderDTO getOrderById(Long id);
-	public void shipOrder(Long id);
-	public void finishOrder(Long id);
-	public void cancelOrder(Long id);
-}
